Move section title
